requirements.txt
asgiref==3.5.2
cffi==1.15.1
cryptography==38.0.3
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.1.3
django-admin-volt==1.0.3
django-cleanup==6.0.0
django-environ==0.9.0
django-ninja==0.19.1
django-resized==1.0.2
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.2
et-xmlfile==1.1.0
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.10
Pillow==9.3.0
pycparser==2.21
pydantic==1.10.2
PyJWT==1.7.1
PyMySQL==1.0.2
pytz==2022.6
PyYAML==6.0
six==1.16.0
sqlparse==0.4.3
tablib==3.2.1
transliterate==1.10.2
typing_extensions==4.4.0
xlrd==2.0.1
xlwt==1.3.0
